fix: separate Tabs and TabHeader
